added ana_SS_rates.py to calculate the rates from a hdWE steady state run. added ana_plainMD.py to calculate PMF and rates between two states from a plain MD run. removed obsolete ana_plainMD_calculatePMF.py

+++ initiate.py
@@ -64,7 +64,7 @@ def parseInitialBoundaries(config):
     for boundary_string in config_string.split(','):
         initial_boundaries.append([])
         for value in boundary_string.split():
-            initial_boundaries[-1].append(int(value))
+            initial_boundaries[-1].append(float(value))
     return initial_boundaries
 
 def parseState(config_string):
